backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
Django==3.1
django-cors-headers==3.5.0
django-seed==0.2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-jwt==1.17.0
drf-yasg==1.17.1
factory-boy==3.0.1
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
